use the svg version of the logo
The PDF generated from the SVG is used. If this is not available / does not work, the PNG file is automatically used. The SVG is not used directly for the case that users of the framework do not have inkscape installed. If users still want to edit or replace the SVG file, it is included and the corresponding command can be commented in.